requirements.txt
ccxt==4.4.80
pandas==2.2.2
numpy==1.26.4
tenacity==8.5.0
pydantic==2.8.2
python-dateutil==2.9.0.post0
PyYAML==6.0.2
optuna==3.6.1
requests==2.31.0
